requirements.txt
altgraph==0.17
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
future==0.18.2
idna==2.10
kiwisolver==1.2.0
matplotlib==3.2.2
mysqlclient==1.4.6
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
PyInstaller==3.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
tkintertable==1.3.2
urllib3==1.25.9
